+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add: Translation
 - Fix: #67 HTML image tags are not recognized for `image` option
 - Add: Option to disable automatic update of a note overview #57
+- Add: Option to configure the `status` field per note overview #66 @aiosk
 
 ## v1.6.0 (2022-12-26)